free-bdr-pipeline/requirements.txt
streamlit
playwright
PyYAML
requests
gspread
oauth2client
